Back to using Double, which Gemini assures me is a beloved choice in the tradeoff space.

+++ ProgressionPlayer/Sources/Tones/Arrow.swift
@@ -8,7 +8,7 @@
 import Accelerate
 import AVFAudio
 
-typealias CoreFloat = Float
+typealias CoreFloat = Double
 let MAX_BUFFER_SIZE = 4096
 
 class Arrow11 {
